pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "archiveinator"
version = "0.1.0"
description = "A local, self-hosted web page archiver with ad blocking and paywall bypass"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40.0",
    "adblock>=0.6.0",
    "trafilatura>=1.8.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
    "platformdirs>=4.0.0",
]

[project.scripts]
archiveinator = "archiveinator.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpserver>=1.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-PyYAML>=6.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["archiveinator"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "e2e: end-to-end tests requiring network access (deselect with '-m not e2e')",
]
